Fix $anon_vote option.

+++ includes/post_vote.php
@@ -15,20 +15,31 @@
 		$query = "SELECT COUNT(*) FROM $post_vote_table WHERE post_id='$id' AND ip='$ip'".$query_part;
 		$result = $db->query($query);
 		$row = $result->fetch_assoc();
+
+		if (!$anon_vote && $user_id == 0)
+		{
+			echo "Anonymous ratings are disabled.";
+			exit;
+		}
 		if($row['COUNT(*)'] < 1)
 		{
 			$result->free_result();
-			if($type == "up")
+			if($type == "up") {
 				$query = "UPDATE $post_table SET score=score+1 WHERE id='$id'";
-			else if($type == "down")
+				$db->query($query);
+			} else if($type == "down") {
 				$query = "UPDATE $post_table SET score=score-1 WHERE id='$id'";
-			else
+				$db->query($query);
+			} else
 				exit;
-			$db->query($query);
 			$query = "INSERT INTO $post_vote_table(rated, ip, post_id, user_id) VALUES('$type', '$ip', '$id', '$user_id')";
 			$db->query($query);
 			$cache = new cache();
 			$cache->destroy("cache/$id/post.cache");
+		}
+		else
+		{
+			$result->free_result();
 		}
 		$query = "SELECT score FROM $post_table WHERE id='$id'";
 		$result = $db->query($query);
